Combine both ColorExtensions files
Oops, there were accidentally two
