--- Assets/Scripts/Colour/ColorExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3b11c40aea98c04d87def566a8114b4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
