+++ Assets/Scripts/Colour/Compositing/AlphaCompositing.cs
@@ -1,3 +1,5 @@
+using PAC.Extensions;
+
 using UnityEngine;
 
 namespace PAC.Colour.Compositing
